Conflicts:
	CvGameCoreDLL_Expansion2/CvAIOperation.cpp
	CvGameCoreDLL_Expansion2/CvEconomicAI.cpp
	CvGameCoreDLL_Expansion2/CvGameCoreDLL_Expansion2.vs2010.vcxproj
	CvGameCoreDLL_Expansion2/CvHomelandAI.cpp
	CvGameCoreDLL_Expansion2/CvHomelandAI.h
	CvGameCoreDLL_Expansion2/CvPlayer.cpp
	CvGameCoreDLL_Expansion2/CvTacticalAI.cpp

+++ CvGameCoreDLL_Expansion2/CvMilitaryAI.cpp
@@ -549,8 +549,6 @@ bool CvMilitaryAI::RequestSneakAttack(PlayerTypes eEnemy)
 	CvAIOperation* pOperation = 0;
 	int iOperationID;
 	// Let's only allow us to be sneak attacking one opponent at a time, so abort if already have one of these operations active against any opponent
-#if defined(MOD_BALANCE_CORE_MILITARY)
-#else
 	if (m_pPlayer->haveAIOperationOfType(AI_OPERATION_NAVAL_SNEAK_ATTACK, &iOperationID))
 	{
 		return false;
@@ -559,7 +557,6 @@ bool CvMilitaryAI::RequestSneakAttack(PlayerTypes eEnemy)
 	{
 		return false;
 	}
-#endif
 #if defined(MOD_BALANCE_CORE_MILITARY)
 	target = FindBestAttackTarget2(AI_OPERATION_SNEAK_CITY_ATTACK, eEnemy);
 #else
@@ -570,12 +567,6 @@ bool CvMilitaryAI::RequestSneakAttack(PlayerTypes eEnemy)
 	{
 		if(target.m_bAttackBySea)
 		{
-#if defined(MOD_BALANCE_CORE_MILITARY)
-			if (m_pPlayer->haveAIOperationOfType(AI_OPERATION_NAVAL_SNEAK_ATTACK, &iOperationID))
-			{
-				return false;
-			}
-#endif
 			if(IsAttackReady(MUFORMATION_NAVAL_INVASION, AI_OPERATION_SNEAK_CITY_ATTACK))
 			{
 				pOperation = m_pPlayer->addAIOperation(AI_OPERATION_NAVAL_SNEAK_ATTACK, eEnemy, target.m_pTargetCity->getArea(), target.m_pTargetCity, target.m_pMusterCity);
@@ -592,12 +583,6 @@ bool CvMilitaryAI::RequestSneakAttack(PlayerTypes eEnemy)
 		}
 		else
 		{
-#if defined(MOD_BALANCE_CORE_MILITARY)
-			if (m_pPlayer->haveAIOperationOfType(AI_OPERATION_SNEAK_CITY_ATTACK, &iOperationID))
-			{
-				return false;
-			}
-#endif
 			if(IsAttackReady((GC.getGame().getHandicapInfo().GetID() > 4 && !(GC.getMap().GetAIMapHint() & 1)) ? MUFORMATION_BIGGER_CITY_ATTACK_FORCE : MUFORMATION_BASIC_CITY_ATTACK_FORCE, AI_OPERATION_SNEAK_CITY_ATTACK))
 			{
 				pOperation = m_pPlayer->addAIOperation(AI_OPERATION_SNEAK_CITY_ATTACK, eEnemy, target.m_pTargetCity->getArea(), target.m_pTargetCity, target.m_pMusterCity);
@@ -2773,7 +2758,11 @@ int CvMilitaryAI::GetBarbarianThreatTotal()
 	ScanForBarbarians();
 
 	// Major threat for each camp seen
+#if defined(MOD_BALANCE_CORE)
+	iRtnValue += GC.getAI_MILITARY_THREAT_WEIGHT_SEVERE() * m_iBarbarianCampCount;
+#else
 	iRtnValue += GC.getAI_MILITARY_THREAT_WEIGHT_MAJOR() * m_iBarbarianCampCount;
+#endif
 
 	// One minor threat for every X barbarians
 	iRtnValue += m_iVisibleBarbarianCount / GC.getAI_MILITARY_BARBARIANS_FOR_MINOR_THREAT();
@@ -2806,10 +2795,16 @@ int CvMilitaryAI::GetThreatWeight(ThreatTypes eThreat)
 }
 
 /// How many civs are we fighting?
+#if defined(MOD_BALANCE_CORE)
+int CvMilitaryAI::GetNumberCivsAtWarWith(bool bMinor) const
+#else
 int CvMilitaryAI::GetNumberCivsAtWarWith() const
+#endif
 {
 	PlayerTypes eLoopPlayer;
+#if !defined(MOD_BALANCE_CORE)
 	WarStateTypes eWarState;
+#endif
 	int iRtnValue = 0;
 
 	// Let's figure out if we're at war
@@ -2822,11 +2817,22 @@ int CvMilitaryAI::GetNumberCivsAtWarWith() const
 		eLoopPlayer = (PlayerTypes) iPlayerLoop;
 
 		// Is this a player we have relations with?
+#if defined(MOD_BALANCE_CORE)
+		if(GET_PLAYER(eLoopPlayer).isMinorCiv() && !bMinor)
+		{
+			continue;
+		}
+		if(eLoopPlayer != m_pPlayer->GetID())
+		{
+			if(GET_TEAM(m_pPlayer->getTeam()).isAtWar(GET_PLAYER(eLoopPlayer).getTeam()))
+			{
+#else
 		if(eLoopPlayer != m_pPlayer->GetID() && m_pPlayer->GetDiplomacyAI()->IsPlayerValid(eLoopPlayer))
 		{
 			eWarState = m_pPlayer->GetDiplomacyAI()->GetWarState(eLoopPlayer);
 			if(eWarState != NO_WAR_STATE_TYPE)
 			{
+#endif
 				iRtnValue++;
 			}
 		}
@@ -3980,6 +3986,12 @@ void CvMilitaryAI::UpdateMilitaryStrategies()
 void CvMilitaryAI::UpdateOperations()
 {
 #if defined(MOD_BALANCE_CORE_MILITARY)
+#if defined(MOD_BALANCE_CORE)
+	if(m_pPlayer->isMinorCiv() || m_pPlayer->isBarbarian())
+	{
+		return;
+	}
+#endif
 	AI_PERF_FORMAT("Military-AI-perf.csv", ("UpdateOperations, Turn %03d, %s", GC.getGame().getElapsedGameTurns(), m_pPlayer->getCivilizationShortDescription()) );
 
 	int iOperationID;
@@ -4053,14 +4065,37 @@ void CvMilitaryAI::UpdateOperations()
 	EconomicAIStrategyTypes eStrategyExpandToOtherContinents = (EconomicAIStrategyTypes) GC.getInfoTypeForString("ECONOMICAISTRATEGY_EXPAND_TO_OTHER_CONTINENTS");
 	if(m_pPlayer->GetEconomicAI()->IsUsingStrategy(eStrategyExpandToOtherContinents))
 	{
-		bWillingToAcceptRisk = true;
+		if((m_iTotalThreatWeight / 4) < GetBarbarianThreatTotal())
+		{
+			bWillingToAcceptRisk = true;
+		}
+	}
+	EconomicAIStrategyTypes eNeedRecon = (EconomicAIStrategyTypes) GC.getInfoTypeForString("ECONOMICAISTRATEGY_NEED_RECON");
+	if(m_pPlayer->GetEconomicAI()->IsUsingStrategy(eNeedRecon))
+	{
+		if((m_iTotalThreatWeight / 4) < GetBarbarianThreatTotal())
+		{
+			bWillingToAcceptRisk = true;
+		}
+	}
+	EconomicAIStrategyTypes eEarlyExpansion = (EconomicAIStrategyTypes) GC.getInfoTypeForString("ECONOMICAISTRATEGY_EARLY_EXPANSION");
+	if(m_pPlayer->GetEconomicAI()->IsUsingStrategy(eEarlyExpansion))
+	{
+		if((m_iTotalThreatWeight / 4) < GetBarbarianThreatTotal())
+		{
+			bWillingToAcceptRisk = true;
+		}
 	}
 #endif
 	//
 	// Operations vs. Barbarians
 	//
 	// If have aborted the eradicate barbarian strategy or if the threat level from civs is significantly higher than from barbs, we better abort all of them
+#if defined(MOD_BALANCE_CORE)
+	if(!IsUsingStrategy(eStrategyBarbs) || (IsUsingStrategy(eStrategyFightAWar) && !bWillingToAcceptRisk))
+#else
 	if(!IsUsingStrategy(eStrategyBarbs) || IsUsingStrategy(eStrategyFightAWar) || !bWillingToAcceptRisk)
+#endif
 	{
 		bool bFoundOneToDelete = true;
 		while(bFoundOneToDelete)
@@ -4266,18 +4301,24 @@ void CvMilitaryAI::UpdateOperations()
 					}
 					if(m_pPlayer->haveAIOperationOfType(AI_OPERATION_CITY_STATE_NAVAL_ATTACK, &iOperationID, eLoopPlayer))
 					{
-						pOperation = m_pPlayer->getAIOperation(iOperationID);
-						if(pOperation)
+						if(GET_PLAYER(eLoopPlayer).isMinorCiv() && GET_PLAYER(eLoopPlayer).GetMinorCivAI()->GetAlly() != NO_PLAYER)
 						{
-							pOperation->Kill(AI_ABORT_WAR_STATE_CHANGE);
+							pOperation = m_pPlayer->getAIOperation(iOperationID);
+							if(pOperation)
+							{
+								pOperation->Kill(AI_ABORT_WAR_STATE_CHANGE);
+							}
 						}
 					}
 					if(m_pPlayer->haveAIOperationOfType(AI_OPERATION_CITY_STATE_ATTACK, &iOperationID, eLoopPlayer))
 					{
-						pOperation = m_pPlayer->getAIOperation(iOperationID);
-						if(pOperation)
+						if(GET_PLAYER(eLoopPlayer).isMinorCiv() && GET_PLAYER(eLoopPlayer).GetMinorCivAI()->GetAlly() != NO_PLAYER)
 						{
-							pOperation->Kill(AI_ABORT_WAR_STATE_CHANGE);
+							pOperation = m_pPlayer->getAIOperation(iOperationID);
+							if(pOperation)
+							{
+								pOperation->Kill(AI_ABORT_WAR_STATE_CHANGE);
+							}
 						}
 					}
 					break;
@@ -4432,8 +4473,8 @@ void CvMilitaryAI::UpdateOperations()
 							{
 								iFilledSlots = MilitaryAIHelpers::NumberOfFillableSlots(m_pPlayer, MUFORMATION_CLOSE_CITY_DEFENSE, false, &iNumRequiredSlots);
 
-								// Not willing to build units to get this off the ground
-								if (iFilledSlots >= iNumRequiredSlots)
+								// Gogogo!
+								if (iFilledSlots > 0)
 								{
 									m_pPlayer->addAIOperation(AI_OPERATION_CITY_CLOSE_DEFENSE, NO_PLAYER, -1, pTarget, pTarget);
 								}
@@ -4447,8 +4488,8 @@ void CvMilitaryAI::UpdateOperations()
 							{
 								iFilledSlots = MilitaryAIHelpers::NumberOfFillableSlots(m_pPlayer, MUFORMATION_CLOSE_CITY_DEFENSE, false, &iNumRequiredSlots);
 
-								// Not willing to build units to get this off the ground
-								if (iFilledSlots >= iNumRequiredSlots)
+								//Gogogogo....lightly.
+								if (iFilledSlots > (iNumRequiredSlots / 2))
 								{
 									m_pPlayer->addAIOperation(AI_OPERATION_CITY_CLOSE_DEFENSE, NO_PLAYER, -1, pTarget, pTarget);
 								}
@@ -4654,7 +4695,7 @@ void CvMilitaryAI::UpdateOperations()
 						}
 					}
 				}
-				else if(GET_PLAYER(eLoopPlayer).isMinorCiv() && eWarState > WAR_STATE_DEFENSIVE)
+				else if(GET_PLAYER(eLoopPlayer).isMinorCiv() && eWarState >= WAR_STATE_DEFENSIVE)
 				{
 					target = FindBestAttackTarget2(AI_OPERATION_CITY_STATE_NAVAL_ATTACK, eLoopPlayer, &iScore);
 					if(target.m_pTargetCity && target.m_bAttackBySea)
@@ -5371,7 +5412,11 @@ void CvMilitaryAI::DisbandObsoleteUnits()
 	int iLandScore = MAX_INT;
 
 	// Don't do this if at war
+#if defined(MOD_BALANCE_CORE)
+	if(GetNumberCivsAtWarWith(false) > 0)
+#else
 	if(GetNumberCivsAtWarWith() > 0)
+#endif
 	{
 		return;
 	}
@@ -5510,7 +5555,15 @@ UnitHandle CvMilitaryAI::FindBestUnitToScrap(bool bLand, bool bDeficitForcedDisb
 
 			// Can I still build this unit? If so too new to scrap
 			if(bLand && m_pPlayer->canTrain(pLoopUnit->getUnitType(), false /*bContinue*/, true /*bTestVisible*/, true /*bIgnoreCost*/))
-				bStillNeeded = true;
+				//But not for scouts - let's pick those off first.
+				if(bLand && pLoopUnit->AI_getUnitAIType() != UNITAI_EXPLORE)
+				{
+					bStillNeeded = true;
+				}
+				else if(!bLand && pLoopUnit->AI_getUnitAIType() != UNITAI_EXPLORE_SEA)
+				{
+					bStillNeeded = true;
+				}
 
 			// Is this a ship on a water body without enemies?
 			if (!bLand)
@@ -6538,14 +6591,44 @@ void CvMilitaryAI::MinorAttackTest()
 				{
 					if(GET_PLAYER(eLoopPlayer).GetProximityToPlayer(m_pPlayer->GetID()) >= PLAYER_PROXIMITY_CLOSE)
 					{
-#if defined(MOD_BALANCE_CORE_MILITARY)
-						target = FindBestAttackTarget2(AI_OPERATION_SMALL_CITY_ATTACK, eLoopPlayer);
-#else
-						target = FindBestAttackTarget(AI_OPERATION_SMALL_CITY_ATTACK, eLoopPlayer);
-#endif
-						if(target.m_pTargetCity)
+						bool bHasOperationUnderway = false;
+						bHasOperationUnderway = m_pPlayer->haveAIOperationOfType(AI_OPERATION_NAVAL_BOMBARDMENT, &iOperationID);
+						bHasOperationUnderway = m_pPlayer->haveAIOperationOfType(AI_OPERATION_PILLAGE_ENEMY, &iOperationID);
+						if (!bHasOperationUnderway)
 						{
-							RequestSpecificAttack(target, 1);
+							target = FindBestAttackTarget2(AI_OPERATION_SMALL_CITY_ATTACK, eLoopPlayer);
+							if(target.m_pTargetCity && target.m_bAttackBySea && m_pPlayer->getCapitalCity()->isCoastal())
+							{
+								bool bHasOperationUnderway = m_pPlayer->haveAIOperationOfType(AI_OPERATION_NAVAL_BOMBARDMENT, &iOperationID);
+								if (!bHasOperationUnderway)
+								{
+									int iNumRequiredSlots;
+									int iFilledSlots = MilitaryAIHelpers::NumberOfFillableSlots(m_pPlayer, MUFORMATION_NAVAL_BOMBARDMENT, true, &iNumRequiredSlots);
+
+									// Not willing to build units to get this off the ground
+									if(iFilledSlots >= iNumRequiredSlots)
+									{
+										m_pPlayer->addAIOperation(AI_OPERATION_NAVAL_BOMBARDMENT, eLoopPlayer);
+										break;
+									}
+								}
+							}
+							else if(target.m_pTargetCity)
+							{
+								bool bHasOperationUnderway = m_pPlayer->haveAIOperationOfType(AI_OPERATION_PILLAGE_ENEMY, &iOperationID);
+								if (!bHasOperationUnderway)
+								{
+									int iNumRequiredSlots;
+									int iFilledSlots = MilitaryAIHelpers::NumberOfFillableSlots(m_pPlayer, MUFORMATION_FAST_PILLAGERS, false, &iNumRequiredSlots);
+
+									// Not willing to build units to get this off the ground
+									if(iFilledSlots >= iNumRequiredSlots)
+									{
+										m_pPlayer->addAIOperation(AI_OPERATION_PILLAGE_ENEMY, eLoopPlayer);
+										break;
+									}
+								}
+							}
 						}
 					}
 				}
@@ -6765,9 +6848,24 @@ bool MilitaryAIHelpers::IsTestStrategy_WarMobilization(MilitaryAIStrategyTypes e
 }
 
 /// "At War" Player Strategy: If the player is at war, increase OFFENSE, DEFENSE and MILITARY_TRAINING.  Then look into which operation(s) to run
+#if defined(MOD_BALANCE_CORE)
+bool MilitaryAIHelpers::IsTestStrategy_AtWar(CvPlayer* pPlayer, bool bMinor)
+#else
 bool MilitaryAIHelpers::IsTestStrategy_AtWar(CvPlayer* pPlayer)
+#endif
 {
+#if defined(MOD_BALANCE_CORE)
+	if(pPlayer->GetMilitaryAI()->GetNumberCivsAtWarWith(bMinor) > 0)
+	{
+		return true;
+	}
+	else
+	{
+		return false;
+	}
+#else
 	return (pPlayer->GetMilitaryAI()->GetNumberCivsAtWarWith() > 0);
+#endif
 }
 
 /// "Minor Civ GeneralDefense" Player Strategy: Prioritize CITY_DEFENSE and DEFENSE
@@ -7315,6 +7413,37 @@ int MilitaryAIHelpers::NumberOfFillableSlots(CvPlayer* pPlayer, MultiunitFormati
 			}
 		}
 	}
+#if defined(MOD_BALANCE_CORE)
+	CvCity* pLoopCity;
+	int iLoop2;
+	for(pLoopCity = pPlayer->firstCity(&iLoop2); pLoopCity != NULL; pLoopCity = pPlayer->nextCity(&iLoop2))
+	{
+		if(pLoopCity != NULL)
+		{
+			if(pLoopCity->isProductionUnit())
+			{
+				UnitTypes eUnit = pLoopCity->getProductionUnit();
+				if(eUnit != NO_UNIT)
+				{
+					CvUnitEntry* pkUnitEntry = GC.getUnitInfo(eUnit);
+					if(pkUnitEntry && pLoopCity->getProductionTurnsLeft(eUnit, 0) <= 5)
+					{
+						for(it = slotsToFill.begin(); it != slotsToFill.end(); it++)
+						{
+							CvFormationSlotEntry slotEntry = *it;
+							if(pkUnitEntry->GetUnitAIType((UnitAITypes)slotEntry.m_primaryUnitType) || pkUnitEntry->GetUnitAIType((UnitAITypes)slotEntry.m_secondaryUnitType))
+							{
+								slotsToFill.erase(it);
+								iWillBeFilled++;
+								break;
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+#endif
 
 	// Now go back through remaining slots and see how many were required, we'll need that many more units
 	if(piNumberSlotsRequired != NULL)
